Half of the job done! (or is it...?) - many bugs, and unclear things

+++ src/implementation/MyCode.java
@@ -3,13 +3,43 @@ package implementation;
 import code.GuiException;
 
 import java.io.*;
+import java.math.BigInteger;
 import java.security.*;
 import java.security.cert.*;
 import java.security.cert.Certificate;
-import java.util.Enumeration;
+import java.security.interfaces.DSAPublicKey;
+import java.security.spec.DSAPublicKeySpec;
+import java.security.spec.ECPublicKeySpec;
+import java.security.spec.RSAPublicKeySpec;
+import java.text.DateFormat;
+import java.util.*;
 
+import com.sun.xml.internal.bind.v2.runtime.reflect.opt.Const;
+import gui.Constants;
+import gui.GuiInterfaceV1;
+import org.bouncycastle.asn1.ASN1ObjectIdentifier;
+import org.bouncycastle.asn1.ASN1OctetString;
+import org.bouncycastle.asn1.pkcs.RSAPublicKey;
+import org.bouncycastle.asn1.x500.X500Name;
+import org.bouncycastle.asn1.x500.X500NameBuilder;
+import org.bouncycastle.asn1.x500.style.BCStyle;
+import org.bouncycastle.asn1.x509.*;
+import org.bouncycastle.asn1.x509.Extension;
+import org.bouncycastle.cert.X509v3CertificateBuilder;
+import org.bouncycastle.cert.jcajce.JcaX509CertificateConverter;
+import org.bouncycastle.cert.jcajce.JcaX509v3CertificateBuilder;
+import org.bouncycastle.jce.ECNamedCurveTable;
+import org.bouncycastle.jce.provider.BouncyCastleProvider;
 import org.bouncycastle.openssl.jcajce.JcaPEMWriter;
+import org.bouncycastle.operator.ContentSigner;
+import org.bouncycastle.operator.OperatorCreationException;
+import org.bouncycastle.operator.jcajce.JcaContentSignerBuilder;
 import org.bouncycastle.util.io.pem.PemWriter;
+import sun.security.ec.ECPublicKeyImpl;
+import sun.security.provider.DSAPublicKeyImpl;
+import sun.security.rsa.RSAPublicKeyImpl;
+
+import static gui.GuiInterfaceV1.reportError;
 
 public class MyCode extends x509.v3.CodeV3 {
 
@@ -28,30 +58,27 @@ public class MyCode extends x509.v3.CodeV3 {
             ks.store(fos, pass.toCharArray());
         } catch (KeyStoreException | IOException | NoSuchAlgorithmException | CertificateException e) {
             e.printStackTrace();
-            access.reportError(e);
+            reportError(e);
         }
     }
 
     @Override
     public Enumeration<String> loadLocalKeystore() {
+        Security.addProvider(new BouncyCastleProvider());
         try {
             ks = KeyStore.getInstance("pkcs12");
-            ks.load(null,null);
-            ks.store(new FileOutputStream(file_to_store_ks), pass.toCharArray());
-        } catch (KeyStoreException e) {
+            File fajl = new File(file_to_store_ks);
+            if (fajl.exists()) {
+                FileInputStream fis = new FileInputStream(file_to_store_ks);
+                ks.load(fis, pass.toCharArray());
+                return ks.aliases();
+            } else {
+                ks.load(null, null);
+                ks.store(new FileOutputStream(file_to_store_ks), pass.toCharArray());
+            }
+        } catch (KeyStoreException | IOException | NoSuchAlgorithmException | CertificateException e) {
             e.printStackTrace();
-        } catch (IOException e) {
-            e.printStackTrace();
-        } catch (CertificateException e) {
-            e.printStackTrace();
-        } catch (NoSuchAlgorithmException e) {
-            e.printStackTrace();
-        }
-
-        try {
-            return ks.aliases();
-        } catch (KeyStoreException e) {
-            e.printStackTrace();
+            reportError(e);
         }
 
         return null;
@@ -59,16 +86,233 @@ public class MyCode extends x509.v3.CodeV3 {
 
     @Override
     public void resetLocalKeystore() {
-
+        try {
+            ks.load(null,null);
+        } catch (IOException | NoSuchAlgorithmException | CertificateException e) {
+            reportError(e);
+            e.printStackTrace();
+        }
+        updateKeyStoreFile();
     }
 
     @Override
     public int loadKeypair(String s) {
-        return 0;
+        try {
+            X509Certificate certificate = (X509Certificate) ks.getCertificate(s);
+
+            //SETTING SUBJECT INFO
+            access.setSubject(certificate.getSubjectX500Principal().getName());
+//            Scanner scanner = new Scanner (certificate.getSubjectX500Principal().getName());
+//            scanner.useDelimiter(",");
+//            while (scanner.hasNext()) {
+//                Scanner auxScanner = new Scanner(scanner.next());
+//                auxScanner.useDelimiter("=");
+//                switch (auxScanner.next()){
+//                    case "CN": access.setSubjectCommonName(auxScanner.next());break;
+//                    case "OU": access.setSubjectOrganizationUnit(auxScanner.next()); break;
+//                    case "O": access.setSubjectOrganization(auxScanner.next()); break;
+//                    case "ST": access.setSubjectState(auxScanner.next()); break;
+//                    case "C": access.setSubjectCountry(auxScanner.next()); break;
+//                    case "L": access.setSubjectLocality(auxScanner.next());break;
+//                }
+//            }
+
+            //SETTING CA INFO //TODO proveri ovo
+            access.setIssuer(certificate.getIssuerX500Principal().getName());
+            access.setIssuerSignatureAlgorithm(certificate.getSigAlgName());
+
+            //SETTING CERTIFICATE VALIDITY
+            access.setNotBefore(certificate.getNotBefore());
+            access.setNotAfter(certificate.getNotAfter());
+
+            //CERTIFICATE VERSION
+            access.setVersion(certificate.getVersion() - 1);
+
+            //SERIAL NUMBER
+            access.setSerialNumber(String.valueOf(certificate.getSerialNumber()));
+
+            //SIGNATURE ALGORITHM
+            String alg = certificate.getPublicKey().getAlgorithm();
+            access.setPublicKeyAlgorithm(alg);
+            access.setPublicKeyDigestAlgorithm(certificate.getSigAlgName());
+            if (alg == "RSA") {
+                RSAPublicKeyImpl rsa_alg = (RSAPublicKeyImpl) certificate.getPublicKey();
+                access.setPublicKeyParameter(String.valueOf(rsa_alg.getModulus().bitLength())); //+1 ?
+            }
+
+            if (alg == "DSA") {
+                DSAPublicKeyImpl dsa_alg = (DSAPublicKeyImpl) certificate.getPublicKey();
+                access.setPublicKeyParameter(String.valueOf(dsa_alg.getY().bitLength()));
+            }
+
+            if (alg == "EC") {
+                ECPublicKeyImpl ec_alg = (ECPublicKeyImpl) certificate.getPublicKey();
+                access.setPublicKeyECCurve(String.valueOf(ec_alg.getParams().getCurve()));
+                access.setPublicKeyParameter(String.valueOf(ec_alg.getParams()));
+            }
+
+
+            //VERSION 3 EXTENSIONS
+            if (certificate.getVersion() == 3) {
+
+                //critical
+                access.setCritical(Constants.BC, false);
+                access.setCritical(Constants.SKID, false);
+                access.setCritical(Constants.SDA, false);
+                Set<String> criticalExtensionOIDs = certificate.getCriticalExtensionOIDs();
+                for (String oid : criticalExtensionOIDs) {
+
+                    if (oid == org.bouncycastle.asn1.x509.Extension.basicConstraints.toString()) {
+                        access.setCritical(Constants.BC, true);
+                    }
+
+                    if (oid == org.bouncycastle.asn1.x509.Extension.subjectKeyIdentifier.toString()) {
+                        access.setCritical(Constants.SKID, true);
+                    }
+
+                    if (oid == org.bouncycastle.asn1.x509.Extension.subjectDirectoryAttributes.toString()) {
+                        access.setCritical(Constants.SDA, true);
+                    }
+
+                }
+
+
+                //basic constraints
+                byte[] extensionValue2 = certificate.getExtensionValue("2.5.29.19");
+                if (extensionValue2 != null) {
+                    byte[] subjectOctets2 = ASN1OctetString.getInstance(extensionValue2).getOctets();
+                    BasicConstraints basicConstraints = BasicConstraints.getInstance(subjectOctets2);
+                    access.setPathLen(String.valueOf(basicConstraints.getPathLenConstraint()));
+                    access.setCA(basicConstraints.isCA());
+                }
+
+
+            //subject key identifier
+                byte[] extensionValue = certificate.getExtensionValue("2.5.29.14");
+                if (extensionValue != null) {
+                    byte[] subjectOctets = ASN1OctetString.getInstance(extensionValue).getOctets();
+                    SubjectKeyIdentifier subjectKeyIdentifier = SubjectKeyIdentifier.getInstance(subjectOctets);
+                    access.setSubjectKeyID(String.valueOf(subjectKeyIdentifier.getKeyIdentifier()));
+                    access.setEnabledSubjectKeyID(true);
+                } else {
+                    access.setEnabledSubjectKeyID(false);
+                }
+
+
+            //subject directory attributes
+                byte[] extensionValue1 = certificate.getExtensionValue("2.5.29.9");
+                if (extensionValue1 != null) {
+                    byte[] subjectOctets1 = ASN1OctetString.getInstance(extensionValue1).getOctets();
+                    SubjectDirectoryAttributes subjectDirectoryAttributes = SubjectDirectoryAttributes.getInstance(subjectOctets1);
+
+                    //TODO ne znam sta ispisuje subjectDirectoryAttriutes vektor, pa kad to proverim mozemo popuniti polja
+                    Vector v = subjectDirectoryAttributes.getAttributes();
+//                    for (Object s123 : v) {
+//                        System.out.println(
+//                                s123
+//                        );
+//                    }
+                }
+//                access.setSubjectDirectoryAttribute(Constants.POB, subjectDirectoryAttributes.getAttributes()); //Place of birth
+//                access.setSubjectDirectoryAttribute(Constants.COC, String v); //Country of citizenship
+//               access.setGender();
+//                access.setDateOfBirth();
+
+
+            }
+
+            return 0;
+            //TODO return
+
+        } catch (KeyStoreException e) {
+            e.printStackTrace();
+            reportError(e);
+        }
+
+
+        return -1;
     }
 
     @Override
     public boolean saveKeypair(String s) {
+        try {
+
+            KeyPairGenerator keyPairGenerator = KeyPairGenerator.getInstance("ECDSA", "BC");
+            org.bouncycastle.jce.spec.ECParameterSpec ecParameterSpec = ECNamedCurveTable.getParameterSpec(access.getPublicKeyECCurve());
+            keyPairGenerator.initialize(ecParameterSpec, new SecureRandom());
+
+            //publicKeyInfo
+            KeyPair keyPair = keyPairGenerator.generateKeyPair();
+            PrivateKey privateKey = keyPair.getPrivate();
+            PublicKey subjectPublicKeyInfo = keyPair.getPublic();
+
+            //subject
+            X500NameBuilder x500NameBuilder = new X500NameBuilder();
+            ASN1ObjectIdentifier[] oids = new ASN1ObjectIdentifier[]{BCStyle.C, BCStyle.ST, BCStyle.L, BCStyle.O, BCStyle.OU, BCStyle.CN};
+            String[] values = new String[] {access.getSubjectCountry(), access.getSubjectState(), access.getSubjectLocality(),
+                                            access.getSubjectOrganization(), access.getSubjectOrganizationUnit(), access.getSubjectCommonName()};
+            x500NameBuilder.addMultiValuedRDN(oids,values);
+            X500Name subject = x500NameBuilder.build();
+
+            //issuer
+            X500Name issuer = subject;
+
+            //DateTo DateFrom
+            Date notBefore = access.getNotBefore();
+            Date notAfter = access.getNotAfter();
+
+            //Serial Number
+            BigInteger serialNumber = BigInteger.valueOf(Long.parseLong(access.getSerialNumber()));
+
+
+            X509v3CertificateBuilder certificateBuilder = new JcaX509v3CertificateBuilder(
+                   issuer,
+                   serialNumber,
+                   notBefore,
+                   notAfter,
+                   subject,
+                   subjectPublicKeyInfo
+            );
+
+            if (access.getVersion() == Constants.V3) {
+                access.setCritical(Constants.BC, false);
+                access.setCritical(Constants.SKID, false);
+                access.setCritical(Constants.SDA, false);
+
+                //basic constraints
+                BasicConstraints basicConstraints = new BasicConstraints(Integer.parseInt(access.getPathLen()));
+                certificateBuilder.addExtension(Extension.basicConstraints, access.isCritical(Constants.BC), basicConstraints.getEncoded());
+
+
+                //subject key identifier
+                if (access.getEnabledSubjectKeyID()) {
+                    SubjectKeyIdentifier subjectKeyIdentifier = new SubjectKeyIdentifier(subjectPublicKeyInfo.getEncoded());
+                    certificateBuilder.addExtension(Extension.subjectKeyIdentifier, access.isCritical(Constants.SKID), subjectKeyIdentifier.getEncoded());
+                }
+
+                //subject directory attributes
+                Vector v = new Vector();
+                v.add(access.getSubjectDirectoryAttribute(Constants.POB));
+                v.add(access.getSubjectDirectoryAttribute(Constants.COC));
+                v.add(access.getGender());
+                v.add(access.getDateOfBirth());
+                SubjectDirectoryAttributes subjectDirectoryAttributes = new SubjectDirectoryAttributes(v);
+                certificateBuilder.addExtension(Extension.subjectDirectoryAttributes, access.isCritical(Constants.SDA), subjectDirectoryAttributes.getEncoded());
+            }
+
+            JcaContentSignerBuilder jcaContentSignerBuilder = new JcaContentSignerBuilder(access.getPublicKeyDigestAlgorithm());
+            ContentSigner contentSigner = jcaContentSignerBuilder.build(privateKey);
+
+            X509Certificate certificate = new JcaX509CertificateConverter().getCertificate(certificateBuilder.build(contentSigner));
+
+
+            ks.setKeyEntry(s, privateKey, pass.toCharArray(), new Certificate[]{certificate});
+            updateKeyStoreFile();
+
+        } catch (CertificateException | IOException | InvalidAlgorithmParameterException | NoSuchAlgorithmException | NoSuchProviderException | OperatorCreationException | KeyStoreException e) {
+           e.printStackTrace();
+           reportError(e);
+        }
 
         return false;
     }
@@ -83,7 +327,7 @@ public class MyCode extends x509.v3.CodeV3 {
             return true;
 
         } catch (KeyStoreException e) {
-            access.reportError(e);
+            reportError(e);
             e.printStackTrace();
         }
         return false;
@@ -107,7 +351,7 @@ public class MyCode extends x509.v3.CodeV3 {
             return true;
         } catch (KeyStoreException | IOException | NoSuchAlgorithmException | CertificateException | UnrecoverableKeyException e) {
             e.printStackTrace();
-            access.reportError(e);
+            reportError(e);
         }
         return false;
     }
@@ -124,7 +368,7 @@ public class MyCode extends x509.v3.CodeV3 {
             return true;
         } catch (KeyStoreException | NoSuchAlgorithmException | UnrecoverableKeyException | IOException | CertificateException e) {
             e.printStackTrace();
-            access.reportError(e);
+            reportError(e);
         }
 
         return false;
@@ -143,14 +387,14 @@ public class MyCode extends x509.v3.CodeV3 {
             return true;
         } catch (FileNotFoundException | CertificateException | KeyStoreException e) {
             e.printStackTrace();
-            access.reportError(e);
+            reportError(e);
         } finally {
             if (inStream != null) {
                 try {
                     inStream.close();
                 } catch (IOException e) {
                     e.printStackTrace();
-                    access.reportError(e);
+                    reportError(e);
                 }
             }
         }
@@ -204,6 +448,7 @@ public class MyCode extends x509.v3.CodeV3 {
 
     @Override
     public boolean exportCSR(String s, String s1, String s2) {
+
         return false;
     }
